交易计划变更列表

+++ trade-core/src/main/java/com/centaline/trans/transplan/service/ToTransplanOperateService.java
@@ -23,7 +23,7 @@ public interface ToTransplanOperateService {
 	 * @param ttsTransPlanHistoryBatch
 	 * @return
 	 */
-	public long insertTtsTransPlanHistoryBatch(TtsTransPlanHistoryBatch ttsTransPlanHistoryBatch);
+	public int insertTtsTransPlanHistoryBatch(TtsTransPlanHistoryBatch ttsTransPlanHistoryBatch);
 	
 	/**
 	 * 查限回访跟进信息
